Fixed Botch shake bug, added futureproofing for shake
Made it so that enemies do not shake when a botched spell is cast, and instead when they are damaged (futureproofing for hit/miss, tageting, etc. accomplished by moving spellcasting to pause->cast->pause->update opacity (and death). also adds futureproofing for if we want to include an after pause death animation. Made it so that enemy shake intensity is damage based (not super noticeable yet, we should work on the formula)