Added title sequence
When game is first started, user is prompted to type in a phrase. Upon
typing it, the game transitions to th title screen. Pressing "NEW FILE"
will start the game from the beginning, all other buttons are currently
non-functional.

Since title sequence is long, the user can also start the game directly
from the "MainScene" by opening up the "MainScene" scene, going to the
"StateManager" component in the "StateManager" object, and setting
"start_immediate" to true.