Added TextEvents for effects that occur during dialogue
TextEvents.cs contains all the functions that can be run during text dialogue (currently, only a momentary screen shake). To add a new event, make a function that matches the TextEventDel delegate, put it in the class, and then add the function to the text_event_map Dictionary.

Events are specified in the gameflow dialogue files with square brackets. Format is [event-name=option1,option2,...]